requirements.txt
python-telegram-bot==20.2
hiredis==2.0.0
aioredis==2.0.1